feat: make "about me" more compelling and focused
This commit provides a more focused and compelling representation of my
skills, passion, and problem-solving approach as a software developer.

The previous introduction has been rephrased to make it more concise and
impactful. For concision, it goes from:

> Hi, I'm Cody. I'm a ...

to

> Hi, I'm Cody, a ...

For impact, I shift the focus from what I like doing in a team to the
effect I want to have on a team. I also provide specific examples of
roles and tasks that I enjoy and have this outcome:

> ...a software developer who is passionate about building tooling to
> accelerate development of high quality solutions

to

> ...a software developer with a commitment to efficiency and
> excellence. I'm passionate about enhancing the productivity of those
> around me, whether through platform engineering, refactoring to reduce
> complexity, or any other challenge offered by a full-stack role.

A section following the introduction has been added to emphasize my
problem solving approach:

> Rather than simply addressing immediate challenges, I take pride in
> solving entire classes of problems, seeking to understand their root
> causes and addressing them holistically.

The section listing my familiarity with various languages, frameworks,
testing tools, databases, and services has been condensed into a single
"toolbox" section and updated to better reflect my current abilities:

> Languages: JavaScript, Python, Java, C++
>
> Frameworks: React-Native, Express, React, Svelte
>
> Testing: Detox, Jest, Mocha, Chai, Sinon"
>
> Databases: CouchDB, MongoDB
>
> Services: Auth0, Stripe

to

> My toolbox includes Git, React-Native, TypeScript, Golang, Vue, Hugo,
> CircleCI, GitHub Actions, Fastlane, MySQL, MongoDB, Docker, Auth0, and
> Stripe. I'm also eagerly exploring Rust and WebAssembly as a possible
> front-end replacement for TypeScript.

The concluding statement has been revised for clarity:

> I am not a ____________ programmer, I am a problem solver.

to

> I am more than a programmer; I am a problem solver.